requirements.txt
cloudscraper==1.2.71
beautifulsoup4==4.12.3
lxml==5.2.2
psycopg2-binary==2.9.9
requests==2.31.0
